cpp: move binding examples in examples subdir
