+++ cogip/cpp/CMakeLists.txt
@@ -27,5 +27,4 @@ execute_process(
 find_package(nanobind CONFIG REQUIRED)
 
 # Add C++ extensions
-add_subdirectory(cython_example)
-add_subdirectory(nanobind_example)
+add_subdirectory(examples)
